pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-smart_traffic_env"
version = "0.1.0"
description = "Smart traffic signal control environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "openai>=2.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m smart_traffic_env.server.app
server = "smart_traffic_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["smart_traffic_env", "smart_traffic_env.server"]
package-dir = { "smart_traffic_env" = ".", "smart_traffic_env.server" = "server" }
